Correct bug that was not letting pass the tests.
In particular, this bug was located in the _transform function,
when looping through the test batches to remove the location and
scale effects. This bug was introduced after changing
the global names of the variables.

Also, the keys in bladder_test data have been renamed to follow
the same nomenclature that we now use with respect to the previous one
(Y, b, X..).

Some other typos have been corrected to fulfill code styling.